# flowcatalyst-typescript/apps/platform-frontend/openapi/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    AssignRolesRequest2:
      type: object
      properties:
        roles:
          type: array
          items:
            type: string
...
    CreateServiceAccountRequest:
      type: object
      required:
        - code
        - name
      properties:
        code:
          type: string
          pattern: \S
          maxLength: 50
        name:
          type: string
          pattern: \S
          maxLength: 100
        description:
          type: string
          maxLength: 500
        clientIds:
          type: array
          items:
            type: string
        applicationId:
          type: string
    CreateServiceAccountResponse:
      type: object
      properties:
        serviceAccount:
          $ref: '#/components/schemas/ServiceAccountDto'
        principalId:
          type: string
        oauth:
          $ref: '#/components/schemas/OAuthCredentials'
        webhook:
          $ref: '#/components/schemas/WebhookCredentials'
...
    Instant:
      type: string
      format: date-time
      examples:
        - 2022-03-10T16:15:50Z
...
    OAuthCredentials:
      type: object
      properties:
        clientId:
          type: string
        clientSecret:
          type: string
...
    ServiceAccountDto:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
        clientIds:
          type: array
          items:
            type: string
        applicationId:
          type: string
        active:
          type: boolean
        authType:
          $ref: '#/components/schemas/WebhookAuthType'
        roles:
          type: array
          items:
            type: string
        lastUsedAt:
          $ref: '#/components/schemas/Instant'
        createdAt:
          $ref: '#/components/schemas/Instant'
        updatedAt:
          $ref: '#/components/schemas/Instant'
...
    ServiceAccountListResponse:
      type: object
      properties:
        serviceAccounts:
          type: array
          items:
            $ref: '#/components/schemas/ServiceAccountDto'
        total:
          type: integer
          format: int32
...
    UpdateServiceAccountRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
        clientIds:
          type: array
          items:
            type: string
...
    WebhookAuthType:
      type: string
      enum:
        - NONE
        - BEARER_TOKEN
        - BASIC_AUTH
        - API_KEY
        - HMAC_SIGNATURE
    WebhookCredentials:
      type: object
      properties:
        authToken:
          type: string
        signingSecret:
          type: string
  securitySchemes:
    SecurityScheme:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication
paths:
...
  /api/admin/service-accounts:
    get:
      summary: List service accounts
      description: List all service accounts with optional filters
      tags:
        - BFF - Service Account Admin
      parameters:
        - description: Filter by active status
          name: active
          in: query
          schema:
            type: boolean
        - description: Filter by application ID
          name: applicationId
          in: query
          schema:
            type: string
        - description: Filter by client ID
          name: clientId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of service accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceAccountListResponse'
        '401':
          description: Not authenticated
    post:
      summary: Create a new service account
      description: Creates a service account and returns credentials. Credentials
        are shown only once.
      tags:
        - BFF - Service Account Admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceAccountRequest'
        required: true
      responses:
        '201':
          description: Service account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateServiceAccountResponse'
        '400':
          description: Invalid request or code already exists
...
  /api/admin/service-accounts/{id}:
    put:
      summary: Update service account metadata
      tags:
        - BFF - Service Account Admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServiceAccountRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {}
        '400':
          description: Bad Request
    get:
      summary: Get service account by ID
      tags:
        - BFF - Service Account Admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Service account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceAccountDto'
        '404':
          description: Service account not found
    delete:
      summary: Delete service account
      tags:
        - BFF - Service Account Admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {}
...
  /api/admin/service-accounts/{id}/roles:
    get:
      summary: Get assigned roles
      tags:
        - BFF - Service Account Admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {}
    put:
      summary: Assign roles
      description: Replace all roles with the provided list (declarative assignment)
      tags:
        - BFF - Service Account Admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRolesRequest2'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {}
        '400':
          description: Bad Request
...
info:
  title: FlowCatalyst Platform API
  version: 1.0.0-SNAPSHOT
  description: FlowCatalyst Platform REST API
servers:
  - url: http://localhost:8080
    description: Auto generated value
  - url: http://0.0.0.0:8080
    description: Auto generated value